Generate episode is a function within old monte_carlo algs file
